fix bug of less than 1 region

+++ marker.py
@@ -182,25 +182,26 @@ def click_and_crop(event, x, y, flags, param):
 
     # undo last region (last reset)
     elif event == cv2.EVENT_RBUTTONDOWN:
-        image = read_img(files[file_pos])
-        draw_info(image)
+        if len(regions) > 0:
+            image = read_img(files[file_pos])
+            draw_info(image)
 
-        print('Cleaning last region (with mouse)')
-        regions.pop(-1)
+            print('Cleaning last region (with mouse)')
+            regions.pop(-1)
 
-        filename, file_extension = os.path.splitext(files[file_pos])
-        file_path = files[file_pos].replace(file_extension, ".txt")
-        # file_path = files[file_pos].replace("jpg", "txt")
+            filename, file_extension = os.path.splitext(files[file_pos])
+            file_path = files[file_pos].replace(file_extension, ".txt")
+            # file_path = files[file_pos].replace("jpg", "txt")
 
-        if os.path.isfile(file_path):
-            os.remove(file_path)
+            if os.path.isfile(file_path):
+                os.remove(file_path)
 
-        for region in regions:
-            class_type = region['class']
-            region = region['region']
-            # draw a rectangle around the region of interest
-            cv2.rectangle(image, region[0], region[1], class_colours[class_type], 2)
-            cv2.imshow("image", image)
+            for region in regions:
+                class_type = region['class']
+                region = region['region']
+                # draw a rectangle around the region of interest
+                cv2.rectangle(image, region[0], region[1], class_colours[class_type], 2)
+                cv2.imshow("image", image)
 
 
 
@@ -312,19 +313,19 @@ if __name__ == '__main__':
         
         # if the 'l' key is pressed, reset the last cropping region
         if key == ord("l"):
-            
-            image = read_img(files[file_pos])
-            print('Cleaning last region')
-            regions.pop(-1)
+            if len(regions) > 0:
+                image = read_img(files[file_pos])
+                print('Cleaning last region')
+                regions.pop(-1)
 
-            filename, file_extension = os.path.splitext(files[file_pos])
-            file_path = files[file_pos].replace(file_extension, ".txt")
-            # file_path = files[file_pos].replace("jpg", "txt")
+                filename, file_extension = os.path.splitext(files[file_pos])
+                file_path = files[file_pos].replace(file_extension, ".txt")
+                # file_path = files[file_pos].replace("jpg", "txt")
+                
+                if os.path.isfile(file_path):
+                    os.remove(file_path)
 
-            if os.path.isfile(file_path):
-                os.remove(file_path)
-
-            print_regions()
+                print_regions()
 
         # if the 'q' key is pressed, break from the loop
         if key == ord("q"):
